+++ packages/tool/tasks/index.js
@@ -1,13 +1,17 @@
 'use strict';
 
-const { format }          = require('util');
-const chalk               = require('chalk');
-const command             = require('./command');
-const config              = require('./config');
-const banner              = require('./banner');
-const utils               = require('./utils');
-const cli                 = require('./lib/cli-command-parser');
-const { merge, includes } = require('./lib/misc');
+const { format } = require('util');
+const chalk      = require('chalk');
+const command    = require('./command');
+const config     = require('./config');
+const banner     = require('./banner');
+const utils      = require('./utils');
+const cli        = require('./lib/cli-command-parser');
+const {
+    merge,
+    includes,
+    copy,
+} = require('./lib/misc');
 
 const plugins = cli.loadPlugins();
 
@@ -34,4 +38,5 @@ module.exports = {
     utils,
     merge,
     includes,
+    copy,
 };
